use report info page added
