+++ src/components/AdminPortal/Report/Detailedreport/ReportProjectDetails.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ReportProjectDetails = () => {
+  return (
+    <div>ReportProjectDetails</div>
+  )
+}
+
+export default ReportProjectDetails
