+++ src/components/AdminPortal/Report/Detailedreport/Skillsachieved.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Skillsachieved = () => {
+  return (
+    <div>Skillsachieved</div>
+  )
+}
+
+export default Skillsachieved
