+++ src/components/AdminPortal/Report/Fetcheddataobject.js
@@ -1,0 +1,11 @@
+export const Data = {
+  ID: "useId",
+  FN: "firstName",
+  LN: "lastName",
+  INTID: "internId",
+  SKILLS: "skills",
+  BC: "beginnerCount",
+  IC: "intermediateCount",
+  AC: "advanceCount",
+  D: "durationOfInternship",
+};
