+++ src/components/AdminPortal/Report/Report.css
@@ -31,7 +31,6 @@
   max-width: 1280px;
   min-width: 1280px;
   overflow-x: hidden;
-  overflow-y: hidden;
 }
 .report-header {
   font-family: "Roboto";
@@ -144,6 +143,7 @@
   border: none;
 }
 .data-display-tech {
+  max-height: 18.5rem;
   background-color: #ffffff;
   position: absolute;
   right: 0;
@@ -151,6 +151,8 @@
   box-shadow: 0px 4px 20px rgba(40, 52, 73, 0.15);
   border-radius: 4px;
   padding-left: 0.875rem;
+  border: 1px solid;
+  overflow-y: scroll;
 }
 .report-drop-down-level {
   height: 1.938rem;
